feat: add layout and restructure pages
